use a single AdapterMock definition
