+++ test/unit/mocks/AdapterMock.js
@@ -1,0 +1,12 @@
+class AdapterMock {
+    constructor() {
+        this.metricsList = {
+            BUFFER_STATE: 'BUFFER_STATE'
+        };
+    }
+    getEventsFor() {
+        return null;
+    }
+}
+
+export default AdapterMock;
